// Opale.Front/src/models/Theme.ts
export type Theme =
  | 'light'
  | 'dark'
  | 'spock'
  | 'papillon-light'
  | 'papillon-dark'
  | 'medieval-light'
  | 'medieval-dark';
export type CvdProfile = 'none' | 'protan-deutan' | 'tritan' | 'achromatopsia';
export type VisionProfile = 'normal' | 'low';
export type ReadingProfile = 'normal' | 'dyslexia';